--- settings_windows.go
@@ -1,8 +1,0 @@
-// +build windows
-
-package main
-
-// Flimit defines the max number of watched files
-func (s *Settings) Flimit() error {
-	return nil
-}
